feat: enhance exam show page with improved UI and functionality
- Redesigned the exam show page layout for better user experience.
- Added summary cards for total questions, total points, and subjects used.
- Implemented a responsive header with exam details and action buttons.
- Introduced a preview modal for exams with export options for PDF and DOCX.
- Updated question listing with enhanced metadata display and editing options.
- Added functionality to export exams in different formats.
- Implemented a confirmation dialog for exam deletion.

chore: add routes for subjects, topics, and tags management

- Added resource routes for subjects, topics, and tags to facilitate CRUD operations.

chore: update Tailwind CSS configuration

- Added safelist for grid-cols-4 to ensure proper styling in dynamic components.

+++ database/migrations/2025_12_12_091949_create_questions_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->text('statement');
             $table->text('explanation')->nullable();
             $table->enum('difficulty_level', ['easy', 'medium', 'hard'])->default('medium');
-            $table->integer('points')->default(1);
+            $table->decimal('points', 3, 1)->default(1.0);
             $table->boolean('is_active')->default(true);
             $table->timestamps();
             $table->softDeletes();
